Interface between GRASS geographical information system and R

Interpreted interface between GRASS geographical information system, version 7, and R, based on starting R from within the GRASS environment, or on running R stand-alone and creating a throw-away GRASS environment from within R. The interface uses classes defined in the sp package to hold spatial data.

Details

Index:

readRAST               read GRASS raster files
writeRAST              write GRASS raster files
readVECT               read GRASS vector object files
writeVECT              write GRASS vector object files
gmeta                  read GRASS metadata from the current LOCATION
getLocationProj         return a PROJ.4 string of projection information
gmeta2grd               create a GridTopology object from the GRASS region
vInfo                   return vector geometry information
vColumns                return vector database columns information
vDataCount              return count of vector database rows
vect2neigh              return area neighbours with shared boundary length

Note that the examples now use the smaller subset North Carolina location: https://grass.osgeo.org/sampledata/north_carolina/nc_basic_spm_grass7.tar.gz

Author

Roger Bivand

Maintainer: Roger Bivand <Roger.Bivand@nhh.no>

Examples

use_sp()
run <- FALSE
if (nchar(Sys.getenv("GISRC")) > 0 &&
  read.dcf(Sys.getenv("GISRC"))[1,"LOCATION_NAME"] == "nc_basic_spm_grass7") run <- TRUE
if (run) {
#  require(rgdal)
  elevation <- readRAST("elevation", ignore.stderr=TRUE, plugin=FALSE)
  summary(elevation)
}

#>               Length                Class                 Mode 
#>             60020100 SpatialGridDataFrame                   S4 

if (run) {
  grd <- gmeta2grd(ignore.stderr=TRUE)
  grd
}

#>                       X1     X2
#> cellcentre.offset 596675 185005
#> cellsize              10     10
#> cells.dim           8166   7350

if (run) {
  set.seed(1)
  pts <- sp::spsample(elevation, 200, "random")
  smple <- sp::SpatialPointsDataFrame(pts, data=sp::over(pts, elevation))
  summary(smple)
}

#>                 Length                  Class                   Mode 
#>                    200 SpatialPointsDataFrame                     S4 

if (run) {
  writeVECT(smple, "sp_dem", v.in.ogr_flags=c("overwrite", "o"), ignore.stderr=TRUE)
  bugsDF <- readVECT("schools", ignore.stderr=TRUE, mapset="PERMANENT")
  summary(bugsDF)
}

#> Warning: Discarded datum NAD83_High_Accuracy_Reference_Network in Proj4 definition: +proj=lcc +lat_0=33.75 +lon_0=-79 +lat_1=36.1666666666667 +lat_2=34.3333333333333 +x_0=609601.22 +y_0=0 +ellps=GRS80 +units=m +no_defs

#>                 Length                  Class                   Mode 
#>                    167 SpatialPointsDataFrame                     S4 

if (run) {
  vInfo("streams", ignore.stderr=TRUE)
}

#>      nodes     points      lines boundaries  centroids      areas    islands 
#>       8645          0       8554          0          0          0          0 
#> primitives      map3d 
#>       8554          0 

if (run) {
  vColumns("streams", ignore.stderr=TRUE)
}

#>         storageType       name
#> 1           INTEGER        cat
#> 2  DOUBLE PRECISION     FNODE_
#> 3  DOUBLE PRECISION     TNODE_
#> 4  DOUBLE PRECISION     LPOLY_
#> 5  DOUBLE PRECISION     RPOLY_
#> 6  DOUBLE PRECISION     LENGTH
#> 7  DOUBLE PRECISION FULL_HYDRO
#> 8  DOUBLE PRECISION FULL_HYD_1
#> 9         CHARACTER      FTYPE
#> 10          INTEGER      FCODE
#> 11        CHARACTER       NAME
#> 12        CHARACTER   RCH_CODE
#> 13        CHARACTER     I_vs_P
#> 14        CHARACTER   USGSstrm

if (run) {
  vDataCount("streams", ignore.stderr=TRUE)
}

#> [1] 8554

if (run) {
  streams <- readVECT("streams", type="line",
    remove.duplicates=FALSE, ignore.stderr=TRUE, plugin=FALSE)
  summary(streams)
}

#> Warning: Discarded datum NAD83_High_Accuracy_Reference_Network in Proj4 definition: +proj=lcc +lat_0=33.75 +lon_0=-79 +lat_1=36.1666666666667 +lat_2=34.3333333333333 +x_0=609601.22 +y_0=0 +ellps=GRS80 +units=m +no_defs

#>                Length                 Class                  Mode 
#>                  8554 SpatialLinesDataFrame                    S4